:zap: fix performance

+++ homebrew/install.sh
@@ -1,8 +1,10 @@
 #!/bin/bash
 
-git clone https://github.com/Homebrew/brew ~/.linuxbrew/Homebrew
-mkdir ~/.linuxbrew/bin
-ln -sfv ~/.linuxbrew/Homebrew/bin/brew ~/.linuxbrew/bin
+if [ ! -f $HOME/.profile ]; then
+  git clone https://github.com/Homebrew/brew ~/.linuxbrew/Homebrew
+  mkdir ~/.linuxbrew/bin
+  ln -sfv ~/.linuxbrew/Homebrew/bin/brew ~/.linuxbrew/bin
+fi
 
 if [ -f $HOME/.profile ]; then
   sed -e '$a export PATH="$HOME/.linuxbrew/bin:$PATH"' $HOME/.profile -i
@@ -15,15 +17,12 @@ export PATH="$HOME/.linuxbrew/bin:$PATH"
 ## Not WSL
 shopt -s nocasematch
 if [ ! "`uname -a | grep 'microsoft'`" ]; then
-  :
-else
   brew install docker
   brew install docker-compose
 fi
 
 brew install git
 brew install gcc
-brew install sudo
 brew install curl
 brew install bat
 brew install exa
